css/css-color/parsing/color-mixed-num-pct.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Color 4: color() parsing, mixed number, percent and none</title>
<link rel="help" href="https://drafts.csswg.org/css-color-4/#color-function">
<link rel="author" title="Chris Lilley" href="mailto:chris@w3.org">
<meta name="assert" content="number, percent and none can be freely intermixed in color function">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="test"></div>
<script>
    const div = document.querySelector("#test");
    function testColorFunction(description, rule, expectedValue) {
        test(function() {
            div.style.color = "black";
            div.style.color = rule;
            assert_equals(getComputedStyle(div).color, expectedValue);
        }, description);
    }

    // Opaque sRGB in color()
    testColorFunction("sRGB all numbers", "color(srgb 1.00 0.50 0.200)", "color(srgb 1 0.5 0.2)");
    testColorFunction("sRGB all percent", "color(srgb 100% 50% 20%)", "color(srgb 1 0.5 0.2)");
    testColorFunction("sRGB mixed number and percent", "color(srgb 100% 0.5 20%)", "color(srgb 1 0.5 0.2)");
    testColorFunction("sRGB mixed number and percent", "color(srgb 1.00 50% 0.2)", "color(srgb 1 0.5 0.2)");
    testColorFunction("sRGB all none", "color(srgb none none none)", "color(srgb 0 0 0)");
    testColorFunction("sRGB number and none", "color(srgb 1.00 none 0.2)", "color(srgb 1 0 0.2)");
    testColorFunction("sRGB percent and none", "color(srgb 100% none 20%)", "color(srgb 1 0 0.2)");
    testColorFunction("sRGB number, percent and none", "color(srgb 100% none 0.2)", "color(srgb 1 0 0.2)");

    // non-unity alpha, sRGB in  color()
    testColorFunction("sRGB with alpha, all numbers", "color(srgb 1.00 0.50 0.200 / 0.6)", "color(srgb 1 0.5 0.2 / 0.6)");
    testColorFunction("sRGB with alpha, all percent", "color(srgb 100% 50% 20%)", "color(srgb 1 0.5 0.2 / 0.6)");
    testColorFunction("sRGB with alpha, mixed number and percent", "color(srgb 100% 0.5 20% / 0.6)", "color(srgb 1 0.5 0.2 / 0.6)");
    testColorFunction("sRGB with alpha, mixed number and percent", "color(srgb 1.00 50% 0.2 / 60%)", "color(srgb 1 0.5 0.2 / 0.6)");
    testColorFunction("sRGB with alpha, all none", "color(srgb none none none / none)", "color(srgb 0 0 0 / 0)");
    testColorFunction("sRGB with alpha, number and none", "color(srgb 1.00 none 0.2 / none)", "color(srgb 1 0 0.2 / 0)");
    testColorFunction("sRGB with alpha, percent and none", "color(srgb 100% none 20% / 30%)", "color(srgb 1 0 0.2 / 0.3)");
    testColorFunction("sRGB with alpha, number, percent and none", "color(srgb 100% none 0.2 / 23.7%)", "color(srgb 1 0 0.2 / 0.237)");

    // Opaque Display P3 in color()
    testColorFunction("Display P3 all numbers", "color(display-p3 1.00 0.50 0.200)", "color(display-p3 1 0.5 0.2)");
    testColorFunction("Display P3 all percent", "color(display-p3 100% 50% 20%)", "color(display-p3 1 0.5 0.2)");
    testColorFunction("Display P3 mixed number and percent", "color(display-p3 100% 0.5 20%)", "color(display-p3 1 0.5 0.2)");
    testColorFunction("Display P3 mixed number and percent", "color(display-p3 1.00 50% 0.2)", "color(display-p3 1 0.5 0.2)");
    testColorFunction("Display P3 all none", "color(display-p3 none none none)", "color(display-p3 0 0 0)");
    testColorFunction("Display P3 number and none", "color(display-p3 1.00 none 0.2)", "color(display-p3 1 0 0.2)");
    testColorFunction("Display P3 percent and none", "color(display-p3 100% none 20%)", "color(display-p3 1 0 0.2)");
    testColorFunction("Display P3 number, percent and none", "color(display-p3 100% none 0.2)", "color(display-p3 1 0 0.2)");

    // non-unity alpha, Display P3 in  color()
    testColorFunction("Display P3 with alpha, all numbers", "color(display-p3 1.00 0.50 0.200 / 0.6)", "color(display-p3 1 0.5 0.2 / 0.6)");
    testColorFunction("Display P3 with alpha, all percent", "color(display-p3 100% 50% 20%)", "color(display-p3 1 0.5 0.2 / 0.6)");
    testColorFunction("Display P3 with alpha, mixed number and percent", "color(display-p3 100% 0.5 20% / 0.6)", "color(display-p3 1 0.5 0.2 / 0.6)");
    testColorFunction("Display P3 with alpha, mixed number and percent", "color(display-p3 1.00 50% 0.2 / 60%)", "color(display-p3 1 0.5 0.2 / 0.6)");
    testColorFunction("Display P3 with alpha, all none", "color(display-p3 none none none / none)", "color(display-p3 0 0 0 / 0)");
    testColorFunction("Display P3 with alpha, number and none", "color(display-p3 1.00 none 0.2 / none)", "color(display-p3 1 0 0.2 / 0)");
    testColorFunction("Display P3 with alpha, percent and none", "color(display-p3 100% none 20% / 30%)", "color(display-p3 1 0 0.2 / 0.3)");
    testColorFunction("Display P3 with alpha, number, percent and none", "color(display-p3 100% none 0.2 / 23.7%)", "color(display-p3 1 0 0.2 / 0.237)");

    // Opaque Rec BT.2020 in  color()
    testColorFunction("Rec BT.2020 all numbers", "color(rec2020 1.00 0.50 0.200)", "color(rec2020 1 0.5 0.2)");
    testColorFunction("Rec BT.2020 all percent", "color(rec2020 100% 50% 20%)", "color(rec2020 1 0.5 0.2)");
    testColorFunction("Rec BT.2020 mixed number and percent", "color(rec2020 100% 0.5 20%)", "color(rec2020 1 0.5 0.2)");
    testColorFunction("Rec BT.2020 mixed number and percent", "color(rec2020 1.00 50% 0.2)", "color(rec2020 1 0.5 0.2)");
    testColorFunction("Rec BT.2020 all none", "color(rec2020 none none none)", "color(rec2020 0 0 0)");
    testColorFunction("Rec BT.2020 number and none", "color(rec2020 1.00 none 0.2)", "color(rec2020 1 0 0.2)");
    testColorFunction("Rec BT.2020 percent and none", "color(rec2020 100% none 20%)", "color(rec2020 1 0 0.2)");
    testColorFunction("Rec BT.2020 number, percent and none", "color(rec2020 100% none 0.2)", "color(rec2020 1 0 0.2)");

    // non-unity alpha, Rec BT.2020 in  color()
    testColorFunction("Rec BT.2020 with alpha, all numbers", "color(rec2020 1.00 0.50 0.200 / 0.6)", "color(rec2020 1 0.5 0.2 / 0.6)");
    testColorFunction("Rec BT.2020 with alpha, all percent", "color(rec2020 100% 50% 20%)", "color(rec2020 1 0.5 0.2 / 0.6)");
    testColorFunction("Rec BT.2020 with alpha, mixed number and percent", "color(rec2020 100% 0.5 20% / 0.6)", "color(rec2020 1 0.5 0.2 / 0.6)");
    testColorFunction("Rec BT.2020 with alpha, mixed number and percent", "color(rec2020 1.00 50% 0.2 / 60%)", "color(rec2020 1 0.5 0.2 / 0.6)");
    testColorFunction("Rec BT.2020 with alpha, all none", "color(rec2020 none none none / none)", "color(rec2020 0 0 0 / 0)");
    testColorFunction("Rec BT.2020 with alpha, number and none", "color(rec2020 1.00 none 0.2 / none)", "color(rec2020 1 0 0.2 / 0)");
    testColorFunction("Rec BT.2020 with alpha, percent and none", "color(rec2020 100% none 20% / 30%)", "color(rec2020 1 0 0.2 / 0.3)");
    testColorFunction("Rec BT.2020 with alpha, number, percent and none", "color(rec2020 100% none 0.2 / 23.7%)", "color(rec2020 1 0 0.2 / 0.237)");

</script>
